Refactoring config code
