+++ PCApp/CameraSlider.UI/MainWindow.xaml.cs
@@ -6,250 +6,13 @@ using System.Windows.Input;
 using System.Diagnostics;
 using CameraSlider.Bluetooth.Events;
 using System.ComponentModel;
-using Color = System.Drawing.Color;
 using CameraSlider.Bluetooth;
 using System.Threading;
-using System.Configuration;
 using System.Text.RegularExpressions;
-using System.Collections.Generic;
-using Newtonsoft.Json;
-using System.Net;
-using System.IO;
-using NHttp;
-using System.Linq;
+using CameraSlider.UI.Config;
 
 namespace CameraSlider.UI
 {
-	public class PresetSettings
-	{
-		public string PresetName
-		{
-			get; set;
-		}
-		public float SlidePosition
-		{
-			get; set;
-		}
-		public float PanPosition
-		{
-			get; set;
-		}
-		public float TiltPosition
-		{
-			get; set;
-		}
-
-		public PresetSettings()
-		{
-			PresetName = "";
-			SlidePosition = 0;
-			PanPosition = 0;
-			TiltPosition = 0;
-		}
-
-		public PresetSettings(PresetSettings other)
-		{
-			PresetName = other.PresetName;
-			SlidePosition = other.SlidePosition;
-			PanPosition = other.PanPosition;
-			TiltPosition = other.TiltPosition;
-		}
-	}
-
-	public class CameraSettingsSection : ConfigurationSection
-	{
-		public CameraSettingsSection()
-		{
-			SlidePos = 0.0f;
-			SlideSpeed = 0.0f;
-			SlideAccel = 0.0f;
-			PanPos = 0.0f;
-			PanSpeed = 0.0f;
-			PanAccel = 0.0f;
-			TiltPos = 0.0f;
-			TiltSpeed = 0.0f;
-			TiltAccel = 0.0f;
-			PresetJson = "";
-		}
-
-		[ConfigurationProperty("slide_pos")]
-		public float SlidePos
-		{
-			get
-			{
-				return (float)this["slide_pos"];
-			}
-			set
-			{
-				this["slide_pos"] = value;
-			}
-		}
-
-		[ConfigurationProperty("slide_speed")]
-		public float SlideSpeed
-		{
-			get
-			{
-				return (float)this["slide_speed"];
-			}
-			set
-			{
-				this["slide_speed"] = value;
-			}
-		}
-
-		[ConfigurationProperty("slide_accel")]
-		public float SlideAccel
-		{
-			get
-			{
-				return (float)this["slide_accel"];
-			}
-			set
-			{
-				this["slide_accel"] = value;
-			}
-		}
-
-		[ConfigurationProperty("pan_pos")]
-		public float PanPos
-		{
-			get
-			{
-				return (float)this["pan_pos"];
-			}
-			set
-			{
-				this["pan_pos"] = value;
-			}
-		}
-
-		[ConfigurationProperty("pan_speed")]
-		public float PanSpeed
-		{
-			get
-			{
-				return (float)this["pan_speed"];
-			}
-			set
-			{
-				this["pan_speed"] = value;
-			}
-		}
-
-		[ConfigurationProperty("pan_accel")]
-		public float PanAccel
-		{
-			get
-			{
-				return (float)this["pan_accel"];
-			}
-			set
-			{
-				this["pan_accel"] = value;
-			}
-		}
-
-		[ConfigurationProperty("tilt_pos")]
-		public float TiltPos
-		{
-			get
-			{
-				return (float)this["tilt_pos"];
-			}
-			set
-			{
-				this["tilt_pos"] = value;
-			}
-		}
-
-		[ConfigurationProperty("tilt_speed")]
-		public float TiltSpeed
-		{
-			get
-			{
-				return (float)this["tilt_speed"];
-			}
-			set
-			{
-				this["tilt_speed"] = value;
-			}
-		}
-
-		[ConfigurationProperty("tilt_accel")]
-		public float TiltAccel
-		{
-			get
-			{
-				return (float)this["tilt_accel"];
-			}
-			set
-			{
-				this["tilt_accel"] = value;
-			}
-		}
-
-		[ConfigurationProperty("preset_json")]
-		public string PresetJson
-		{
-			get
-			{
-				return (string)this["preset_json"];
-			}
-			set
-			{
-				this["preset_json"] = value;
-			}
-		}
-	}
-
-	public class ConfigState
-	{
-		public Configuration _configFile;
-		public CameraSettingsSection _cameraSettingsConfig;
-		public List<PresetSettings> _presets = new List<PresetSettings>();
-		public bool _areConfigSettingsDirty = false;
-		public bool _arePresetsDirty = false;
-
-		public void LoadConfig()
-		{
-			// Open App.Config of executable
-			string app_data = Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData);
-
-			ExeConfigurationFileMap fileMap = new ExeConfigurationFileMap();
-			fileMap.ExeConfigFilename = app_data + "\\CameraSliderUI\\CameraSliderUI.config.txt";
-			_configFile = ConfigurationManager.OpenMappedExeConfiguration(fileMap, ConfigurationUserLevel.None);
-
-			_presets = new List<PresetSettings>();
-			_arePresetsDirty = false;
-			_cameraSettingsConfig = (CameraSettingsSection)_configFile.GetSection("camera_settings");
-			if (_cameraSettingsConfig != null)
-			{
-				if (_cameraSettingsConfig.PresetJson != "")
-				{
-					_presets = JsonConvert.DeserializeObject<List<PresetSettings>>(_cameraSettingsConfig.PresetJson);
-				}
-			}
-			else
-			{
-				_cameraSettingsConfig = new CameraSettingsSection();
-				_configFile.Sections.Add("camera_settings", _cameraSettingsConfig);
-			}
-
-			_configFile.Save();
-		}
-
-		public void SaveConfig()
-		{
-			if (_arePresetsDirty)
-			{
-				_cameraSettingsConfig.PresetJson = JsonConvert.SerializeObject(_presets, Formatting.None);
-				_arePresetsDirty = false;
-			}
-			_areConfigSettingsDirty = false;
-			_configFile.Save();
-		}
-	}
 
 	public partial class MainWindow : Window
 	{
